pyproject.toml
[project]
name = "adk-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-adk>=1.0.0",
]
